pyproject.toml
[tool.poetry]
name = "pingpong-study"
version = "6.2"
description = "An app to support the PingPong College RCT study."
authors = ["Evangelos Kassos <github@evangeloskassos.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
sentry-sdk = "^2.50.0"
opentelemetry-api = "^1.20.0"
requests = "^2.32.5"
fastapi = {extras = ["standard"], version = "^0.128.0"}
uvicorn = "^0.40.0"
aiohttp = "^3.13.3"
pyjwt = "^2.10.1"
azure-communication-email = "^1.1.0"
aiosmtplib = "^5.1.0"
opentelemetry-sdk = "^1.23.0"
arrow = "^1.4.0"
pyairtable = "^3.3.0"
click = "^8.3.1"
dotenv = "^0.9.9"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.5.0"
deptry = "^0.24.0"

[tool.deptry]
known_first_party = ["pingpong"]

[tool.poetry.group.local-scripts-qualtrics]
optional = true

[tool.poetry.group.local-scripts-qualtrics.dependencies]
selenium = "^4.40.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
